pyproject.toml
[tool.poetry]
name = "drf-coursework"
version = "0.1.0"
description = ""
authors = ["Khitrov <Hitrov.95@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
coverage = "^7.5.1"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.1"
pillow = "^10.3.0"
django-filter = "^24.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
